requirements.txt
tiktoken>=0.5.1
openai>=1.14.0
pathspec>=0.12.1
PyYAML>=6.0